+++ backend/api/restrictions/delete.go
@@ -17,7 +17,9 @@ import (
 func Delete(ctx *fasthttp.RequestCtx) {
 	var restriction model.Restriction
 	if ok := api.FindByID(ctx.UserValue("id"), &restriction, ctx); ok && restriction.ValidID() {
-		tx := db.Delete(&restriction)
-		api.SuccessJSON(ctx, map[string]interface{}{"success": (tx.Error == nil), "error": tx.Error})
+		if deletable := api.CompareClientID(ctx, restriction.ClientID, "unprocessible"); deletable {
+			tx := db.Delete(&restriction)
+			api.SuccessJSON(ctx, map[string]interface{}{"success": (tx.Error == nil), "error": tx.Error})
+		}
 	}
 }
